requirements.txt
boto==2.46.1
customerio==0.2.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-change-email==0.1.3
django-compressor==2.1.1
django-easysettings==1.1
django-dajax==0.9.2
django-dajaxice==0.7
django-imagekit==4.0
django-mptt==0.8.7
django-storages==1.5.2
django-timezone-field==2.0
djangorestframework==3.6.2
django-rest-swagger==2.1.2
ecdsa==0.13
factory-boy==2.8.1
funcsigs==1.0.2
html5lib==1.0b8
icalendar==3.11.3
jsonfield==2.0.1
Marmir==0.1.3
markdown==2.6.8
mock==2.0.0
pbr==2.0.0
Pillow==4.1.0
psycopg2>=2.7.1
pycrypto==2.6.1
PyPDFLite==0.1.40
python-dateutil==2.6.0
pytz==2017.2
reportlab==3.4.0
social-auth-app-django==1.1.0
sorl-thumbnail>=12.3
stripe==1.53.0
sqlparse==0.2.3
xhtml2pdf==0.0.6
